backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
pydantic==2.7.0
pydantic-settings==2.2.1
google-cloud-vision==3.7.2
google-cloud-firestore==2.16.0
